fix: title and description years should match temporal extent (WIP) TDE-1464
